Refactor: move WeaponClass to `src` module and add type hints
