+++ Launcher.py
@@ -11,7 +11,7 @@ import pygame
 from src.constants import COLORS, PENETRATION_COLORS
 from src.cursor import Cursor
 from src.muzzle_flash import MuzzleFlash
-from src.weapons import Weapon
+from src.weapons import Weapon, WeaponCategory
 
 if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
     BASE_DIR = Path(sys._MEIPASS)
@@ -113,56 +113,6 @@ class Chest(pygame.sprite.Sprite):
         """Opens the chest, unlocks a random weapon, and removes it from the game."""
         self.opened = True
         self.kill()
-
-
-class WeaponCategory:
-    """Represents a category of weapons."""
-
-    def __init__(self, name, weapons):
-        self.name = name
-        self.weapons = weapons
-        self.current_index = self.find_first_unlocked_weapon()
-
-    def find_first_unlocked_weapon(self):
-        """Finds the index of the first unlocked weapon in the category."""
-        for i, weapon in enumerate(self.weapons):
-            if not weapon.locked:
-                return i
-        return None
-
-    def current_weapon(self):
-        """Returns the currently selected weapon."""
-        if self.current_index is not None:
-            return self.weapons[self.current_index]
-        return None
-
-    def next_weapon(self):
-        """Cycles to the next unlocked weapon in the category."""
-        if self.current_index is None:
-            return
-        start_index = self.current_index
-        while True:
-            self.current_index = (self.current_index + 1) % len(self.weapons)
-            if not self.weapons[self.current_index].locked:
-                return
-            if self.current_index == start_index:
-                return
-
-    def previous_weapon(self):
-        """Cycles to the previous unlocked weapon in the category."""
-        if self.current_index is None:
-            return
-        start_index = self.current_index
-        while True:
-            self.current_index = (self.current_index - 1) % len(self.weapons)
-            if not self.weapons[self.current_index].locked:
-                return
-            if self.current_index == start_index:
-                return
-
-    def has_unlocked_weapon(self):
-        """Checks if the category has at least one unlocked weapon."""
-        return any(not weapon.locked for weapon in self.weapons)
 
 
 class Camera:
